requirements.txt
Flask
pymongo
python-dotenv
pdfplumber
PyJWT